Spinning up for the night.
